Add some more tests
