+++ okhttp-brotli/src/main/kotlin/okhttp3/brotli/BrotliInterceptor.kt
@@ -18,6 +18,7 @@ package okhttp3.brotli
 import okhttp3.Interceptor
 import okhttp3.Response
 import okhttp3.ResponseBody.Companion.asResponseBody
+import okhttp3.brotli.internal.uncompress
 import okhttp3.internal.http.promisesBody
 import okio.GzipSource
 import okio.buffer
@@ -31,38 +32,17 @@ import org.brotli.dec.BrotliInputStream
  * responses.  n.b. this replaces the transparent gzip compression in BridgeInterceptor.
  */
 object BrotliInterceptor : Interceptor {
-  override fun intercept(chain: Interceptor.Chain): Response =
-      if (chain.request().header("Accept-Encoding") == null) {
-        val request = chain.request().newBuilder()
-            .header("Accept-Encoding", "br,gzip")
-            .build()
-
-        val response = chain.proceed(request)
-
-        uncompress(response)
-      } else {
-        chain.proceed(chain.request())
-      }
-
-  internal fun uncompress(response: Response): Response {
-    if (!response.promisesBody()) {
-      return response
-    }
-    val body = response.body ?: return response
-    val encoding = response.header("Content-Encoding") ?: return response
-
-    val decompressedSource = when {
-      encoding.equals("br", ignoreCase = true) ->
-        BrotliInputStream(body.source().inputStream()).source().buffer()
-      encoding.equals("gzip", ignoreCase = true) ->
-        GzipSource(body.source()).buffer()
-      else -> return response
-    }
-
-    return response.newBuilder()
-        .removeHeader("Content-Encoding")
-        .removeHeader("Content-Length")
-        .body(decompressedSource.asResponseBody(body.contentType(), -1))
+  override fun intercept(chain: Interceptor.Chain): Response {
+    return if (chain.request().header("Accept-Encoding") == null) {
+      val request = chain.request().newBuilder()
+        .header("Accept-Encoding", "br,gzip")
         .build()
+
+      val response = chain.proceed(request)
+
+      uncompress(response)
+    } else {
+      chain.proceed(chain.request())
+    }
   }
 }
